+++ modules/services/repositories/src/main/java/au/com/shiftyjelly/pocketcasts/repositories/playback/PlaybackService.kt
@@ -15,10 +15,13 @@ import android.support.v4.media.MediaMetadataCompat
 import android.support.v4.media.session.MediaControllerCompat
 import android.support.v4.media.session.MediaSessionCompat
 import android.support.v4.media.session.PlaybackStateCompat
+import android.widget.Toast
 import androidx.media.MediaBrowserServiceCompat
 import au.com.shiftyjelly.pocketcasts.analytics.AnalyticsEvent
 import au.com.shiftyjelly.pocketcasts.analytics.AnalyticsTracker
+import au.com.shiftyjelly.pocketcasts.analytics.SourceView
 import au.com.shiftyjelly.pocketcasts.localization.BuildConfig
+import au.com.shiftyjelly.pocketcasts.localization.R
 import au.com.shiftyjelly.pocketcasts.models.entity.BaseEpisode
 import au.com.shiftyjelly.pocketcasts.models.entity.Podcast
 import au.com.shiftyjelly.pocketcasts.models.entity.PodcastEpisode
@@ -49,16 +52,29 @@ import au.com.shiftyjelly.pocketcasts.utils.Util
 import au.com.shiftyjelly.pocketcasts.utils.log.LogBuffer
 import com.jakewharton.rxrelay2.BehaviorRelay
 import dagger.hilt.android.AndroidEntryPoint
+import io.reactivex.Observable
 import io.reactivex.disposables.CompositeDisposable
+import io.reactivex.disposables.Disposable
 import io.reactivex.rxkotlin.Observables
 import io.reactivex.rxkotlin.addTo
 import io.reactivex.rxkotlin.subscribeBy
+import io.reactivex.schedulers.Schedulers
 import java.util.Timer
 import java.util.TimerTask
+import java.util.concurrent.TimeUnit
 import javax.inject.Inject
 import kotlin.coroutines.CoroutineContext
+import kotlin.time.Duration
+import kotlin.time.Duration.Companion.ZERO
+import kotlin.time.Duration.Companion.seconds
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.flow.catch
+import kotlinx.coroutines.flow.distinctUntilChanged
+import kotlinx.coroutines.flow.flowOn
+import kotlinx.coroutines.flow.launchIn
+import kotlinx.coroutines.flow.map
+import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.rx2.asObservable
 import kotlinx.coroutines.rx2.awaitSingleOrNull
@@ -129,6 +145,8 @@ open class PlaybackService : MediaBrowserServiceCompat(), CoroutineScope {
 
     @Inject lateinit var analyticsTracker: AnalyticsTracker
 
+    @Inject lateinit var sleepTimer: SleepTimer
+
     var mediaController: MediaControllerCompat? = null
         set(value) {
             field = value
@@ -143,6 +161,9 @@ open class PlaybackService : MediaBrowserServiceCompat(), CoroutineScope {
     lateinit var notificationManager: PlayerNotificationManager
 
     private val disposables = CompositeDisposable()
+
+    private var sleepTimerDisposable: Disposable? = null
+    private var currentTimeLeft: Duration = ZERO
 
     override val coroutineContext: CoroutineContext
         get() = Dispatchers.Default
@@ -162,12 +183,15 @@ open class PlaybackService : MediaBrowserServiceCompat(), CoroutineScope {
 
         mediaController = MediaControllerCompat(this, mediaSession)
         notificationManager = PlayerNotificationManagerImpl(this)
+
+        observePlaybackState()
     }
 
     override fun onDestroy() {
         super.onDestroy()
 
         disposables.clear()
+        sleepTimerDisposable?.dispose()
 
         LogBuffer.i(LogBuffer.TAG_PLAYBACK, "Playback service destroyed")
     }
@@ -650,5 +674,68 @@ open class PlaybackService : MediaBrowserServiceCompat(), CoroutineScope {
         val podcasts = (localPodcasts + serverPodcasts).distinctBy { it.uuid }
         // convert podcasts to the media browser format
         return podcasts.mapNotNull { podcast -> convertPodcastToMediaItem(context = this, podcast = podcast, useEpisodeArtwork = settings.artworkConfiguration.value.useEpisodeArtwork) }
+    }
+
+    private fun observePlaybackState() {
+        sleepTimer.stateFlow
+            .map { it }
+            .distinctUntilChanged()
+            .flowOn(Dispatchers.IO)
+            .onEach { state ->
+                onSleepTimerStateChange(state)
+            }
+            .catch { throwable ->
+                Timber.e(throwable, "Error observing SleepTimer state")
+            }
+            .launchIn(this)
+    }
+
+    private fun onSleepTimerStateChange(state: SleepTimerState) {
+        if (state.isSleepTimerRunning && state.timeLeft != ZERO) {
+            startOrUpdateSleepTimer(state.timeLeft)
+        } else {
+            cancelSleepTimer()
+        }
+    }
+
+    private fun startOrUpdateSleepTimer(newTimeLeft: Duration) {
+        if (newTimeLeft == ZERO || newTimeLeft.isNegative()) {
+            return
+        }
+
+        if (sleepTimerDisposable == null || sleepTimerDisposable!!.isDisposed) {
+            currentTimeLeft = newTimeLeft
+
+            sleepTimerDisposable = Observable.interval(1, TimeUnit.SECONDS, Schedulers.computation())
+                .takeWhile { currentTimeLeft > ZERO }
+                .doOnNext {
+                    currentTimeLeft = currentTimeLeft.minus(1.seconds)
+                    sleepTimer.updateSleepTimerStatus(sleepTimeRunning = currentTimeLeft != ZERO, timeLeft = currentTimeLeft)
+
+                    if (currentTimeLeft == 5.seconds) {
+                        playbackManager.performVolumeFadeOut(5.0)
+                    }
+
+                    if (currentTimeLeft <= ZERO) {
+                        LogBuffer.i(LogBuffer.TAG_PLAYBACK, "Paused from sleep timer.")
+                        CoroutineScope(Dispatchers.Main).launch {
+                            Toast.makeText(applicationContext, applicationContext.getString(R.string.player_sleep_timer_stopped_your_podcast), Toast.LENGTH_LONG).show()
+                            playbackManager.restorePlayerVolume()
+                        }
+                        playbackManager.pause(sourceView = SourceView.AUTO_PAUSE)
+                        sleepTimer.updateSleepTimerStatus(sleepTimeRunning = false)
+                        cancelSleepTimer()
+                    }
+                }
+                .subscribe()
+        } else {
+            currentTimeLeft = newTimeLeft
+        }
+    }
+
+    private fun cancelSleepTimer() {
+        sleepTimerDisposable?.dispose()
+        sleepTimerDisposable = null
+        currentTimeLeft = ZERO
     }
 }