+++ modules/services/repositories/src/main/java/au/com/shiftyjelly/pocketcasts/repositories/playback/SleepTimer.kt
@@ -1,28 +1,27 @@
 package au.com.shiftyjelly.pocketcasts.repositories.playback
 
-import android.app.AlarmManager
-import android.app.PendingIntent
-import android.content.Context
-import android.content.Intent
-import android.text.format.DateUtils
 import au.com.shiftyjelly.pocketcasts.analytics.AnalyticsEvent.PLAYER_SLEEP_TIMER_RESTARTED
 import au.com.shiftyjelly.pocketcasts.analytics.AnalyticsTracker
+import au.com.shiftyjelly.pocketcasts.models.entity.BaseEpisode
+import au.com.shiftyjelly.pocketcasts.preferences.Settings
 import au.com.shiftyjelly.pocketcasts.utils.log.LogBuffer
-import dagger.hilt.android.qualifiers.ApplicationContext
-import java.util.Calendar
 import javax.inject.Inject
 import javax.inject.Singleton
 import kotlin.time.Duration
+import kotlin.time.Duration.Companion.ZERO
 import kotlin.time.Duration.Companion.milliseconds
 import kotlin.time.Duration.Companion.minutes
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.update
 
 @Singleton
 class SleepTimer @Inject constructor(
+    private val settings: Settings,
     private val analyticsTracker: AnalyticsTracker,
-    @ApplicationContext private val context: Context,
 ) {
     companion object {
-        private val MIN_TIME_TO_RESTART_SLEEP_TIMER_IN_MINUTES = 5.minutes
+        private val MIN_TIME_TO_RESTART_SLEEP_TIMER_IN_MINUTES = 1.minutes
         private const val TIME_KEY = "time"
         private const val NUMBER_OF_EPISODES_KEY = "number_of_episodes"
         private const val NUMBER_OF_CHAPTERS_KEY = "number_of_chapters"
@@ -31,171 +30,179 @@ class SleepTimer @Inject constructor(
         const val TAG: String = "SleepTimer"
     }
 
-    private var sleepTimeMs: Long? = null
-    private var lastSleepAfterTime: Duration? = null
-    private var lastSleepAfterEndOfChapterTime: Duration? = null
-    private var lastTimeSleepTimeHasFinished: Duration? = null
-    private var lastEpisodeUuidAutomaticEnded: String? = null
+    private var sleepTimerHistory: SleepTimerHistory = SleepTimerHistory.None
 
-    fun sleepAfter(duration: Duration, onSuccess: () -> Unit) {
-        val sleepAt = System.currentTimeMillis().milliseconds + duration
+    private val _stateFlow: MutableStateFlow<SleepTimerState> = MutableStateFlow(SleepTimerState())
+    val stateFlow: StateFlow<SleepTimerState> = _stateFlow
 
-        if (createAlarm(sleepAt.inWholeMilliseconds)) {
-            lastSleepAfterTime = duration
-            cancelAutomaticSleepOnEpisodeEndRestart()
-            cancelAutomaticSleepOnChapterEndRestart()
-            onSuccess()
+    val state: SleepTimerState
+        get() = _stateFlow.value
+
+    fun updateSleepTimerStatus(
+        sleepTimeRunning: Boolean,
+        sleepAfterEpisodes: Int = 0,
+        sleepAfterChapters: Int = 0,
+        timeLeft: Duration? = null,
+    ) {
+        _stateFlow.update { currentState ->
+            currentState.copy(
+                isSleepTimerRunning = sleepTimeRunning,
+                numberOfEpisodesLeft = sleepAfterEpisodes,
+                numberOfChaptersLeft = sleepAfterChapters,
+                timeLeft = if (!sleepTimeRunning || sleepAfterEpisodes != 0 || sleepAfterChapters != 0) {
+                    ZERO
+                } else {
+                    timeLeft ?: currentState.timeLeft
+                },
+            )
         }
     }
 
-    fun addExtraTime(minutes: Int) {
-        val currentTimeMs = sleepTimeMs
-        if (currentTimeMs == null || currentTimeMs < 0) {
+    fun sleepAfter(duration: Duration) {
+        updateSleepTimerStatus(sleepTimeRunning = true, timeLeft = duration)
+
+        sleepTimerHistory = SleepTimerHistory.AfterTime(
+            lastSleepAfterTime = duration,
+            lastFinishedTime = System.currentTimeMillis().milliseconds + duration,
+        )
+    }
+
+    fun addExtraTime(duration: Duration) {
+        val currentTimeLeft: Duration = state.timeLeft
+        if (currentTimeLeft < ZERO) {
             return
         }
-        val time = Calendar.getInstance().apply {
-            timeInMillis = currentTimeMs
-            add(Calendar.MINUTE, minutes)
+        val newTimeLeft = currentTimeLeft + duration
+
+        _stateFlow.update { currentState ->
+            currentState.copy(timeLeft = newTimeLeft)
         }
-        LogBuffer.i(TAG, "Added extra time: $minutes")
-        createAlarm(time.timeInMillis)
+
+        LogBuffer.i(TAG, "Added extra time: $newTimeLeft")
     }
 
-    fun restartTimerIfIsRunning(onSuccess: () -> Unit): Duration? {
-        return if (isSleepAfterTimerRunning) {
-            lastSleepAfterTime?.let { sleepAfter(it, onSuccess) }
-            lastSleepAfterTime
+    /*
+     * This restart only applies if the sleep timer is set to "sleep in x minutes".
+     * Other options like "end of chapter" and "end of episode" do not apply.
+     * */
+    fun restartTimerForSleepAfterTime(): Duration? {
+        return if (state.timeLeft != ZERO && sleepTimerHistory is SleepTimerHistory.AfterTime) {
+            val sleepAfterTime = (sleepTimerHistory as SleepTimerHistory.AfterTime).lastSleepAfterTime
+            sleepAfter(sleepAfterTime)
+            sleepAfterTime
         } else {
             null
         }
     }
 
     fun restartSleepTimerIfApplies(
-        autoSleepTimerEnabled: Boolean,
         currentEpisodeUuid: String,
-        timerState: SleepTimerState,
-        onRestartSleepAfterTime: () -> Unit,
-        onRestartSleepOnEpisodeEnd: () -> Unit,
-        onRestartSleepOnChapterEnd: () -> Unit,
     ) {
-        if (!autoSleepTimerEnabled) return
+        if (!settings.autoSleepTimerRestart.value || sleepTimerHistory is SleepTimerHistory.None) return
+        val lastTimeHasFinished = sleepTimerHistory.lastFinishedTime ?: return
 
-        lastTimeSleepTimeHasFinished?.let { lastTimeHasFinished ->
-            val diffTime = System.currentTimeMillis().milliseconds - lastTimeHasFinished
+        val diffTime = System.currentTimeMillis().milliseconds - lastTimeHasFinished
 
-            if (shouldRestartSleepEndOfChapter(diffTime, timerState.isSleepEndOfChapterRunning)) {
-                onRestartSleepOnChapterEnd()
-                analyticsTracker.track(PLAYER_SLEEP_TIMER_RESTARTED, mapOf(TIME_KEY to END_OF_CHAPTER_VALUE, NUMBER_OF_CHAPTERS_KEY to timerState.numberOfChapters))
-            } else if (shouldRestartSleepEndOfEpisode(diffTime, currentEpisodeUuid, timerState.isSleepEndOfEpisodeRunning)) {
-                onRestartSleepOnEpisodeEnd()
-                analyticsTracker.track(PLAYER_SLEEP_TIMER_RESTARTED, mapOf(TIME_KEY to END_OF_EPISODE_VALUE, NUMBER_OF_EPISODES_KEY to timerState.numberOfEpisodes))
-            } else if (shouldRestartSleepAfterTime(diffTime, timerState.isSleepTimerRunning)) {
-                lastSleepAfterTime?.let {
-                    analyticsTracker.track(PLAYER_SLEEP_TIMER_RESTARTED, mapOf(TIME_KEY to it.inWholeSeconds))
-                    LogBuffer.i(TAG, "Was restarted with ${it.inWholeMinutes} minutes set")
-                    sleepAfter(it, onRestartSleepAfterTime)
+        when (val history = sleepTimerHistory) {
+            is SleepTimerHistory.AfterChapter -> {
+                if (shouldRestartSleepEndOfChapter(diffTime, state.isSleepEndOfChapterRunning)) {
+                    val chapter = settings.getlastSleepEndOfChapter()
+                    LogBuffer.i(TAG, "Sleep timer was restarted with end of $chapter chapter set")
+                    updateSleepTimerStatus(sleepTimeRunning = true, sleepAfterChapters = chapter)
+                    analyticsTracker.track(PLAYER_SLEEP_TIMER_RESTARTED, mapOf(TIME_KEY to END_OF_CHAPTER_VALUE, NUMBER_OF_CHAPTERS_KEY to settings.getlastSleepEndOfChapter()))
                 }
             }
+
+            is SleepTimerHistory.AfterEpisode -> {
+                if (shouldRestartSleepEndOfEpisode(diffTime, currentEpisodeUuid, state.isSleepEndOfEpisodeRunning, history.lastEpisodeUuidAutomaticEnded)) {
+                    val episodes = settings.getlastSleepEndOfEpisodes()
+                    LogBuffer.i(TAG, "Sleep timer was restarted with end of $episodes episodes set")
+                    updateSleepTimerStatus(sleepTimeRunning = true, sleepAfterEpisodes = episodes)
+                    analyticsTracker.track(PLAYER_SLEEP_TIMER_RESTARTED, mapOf(TIME_KEY to END_OF_EPISODE_VALUE, NUMBER_OF_EPISODES_KEY to settings.getlastSleepEndOfEpisodes()))
+                }
+            }
+
+            is SleepTimerHistory.AfterTime -> {
+                if (shouldRestartSleepAfterTime(diffTime, state.isSleepTimerRunning)) {
+                    sleepAfter(history.lastSleepAfterTime)
+                    analyticsTracker.track(PLAYER_SLEEP_TIMER_RESTARTED, mapOf(TIME_KEY to history.lastSleepAfterTime.inWholeSeconds))
+                    LogBuffer.i(TAG, "Was restarted with ${history.lastSleepAfterTime.inWholeMinutes} minutes set")
+                }
+            }
+
+            SleepTimerHistory.None -> {}
         }
     }
 
-    fun setEndOfEpisodeUuid(uuid: String) {
-        LogBuffer.i(TAG, "Episode $uuid was marked as end of episode")
-        lastEpisodeUuidAutomaticEnded = uuid
-        lastTimeSleepTimeHasFinished = System.currentTimeMillis().milliseconds
-        cancelAutomaticSleepAfterTimeRestart()
-        cancelAutomaticSleepOnChapterEndRestart()
+    suspend fun sleepEndOfEpisode(episode: BaseEpisode, onSleepEndOfEpisode: suspend () -> Unit) {
+        if (state.isSleepEndOfEpisodeRunning) {
+            setEndOfEpisodeUuid(episode.uuid)
+            updateSleepTimer {
+                copy(numberOfEpisodesLeft = state.numberOfEpisodesLeft - 1)
+            }
+        }
+
+        if (state.isSleepEndOfEpisodeRunning) return
+
+        updateSleepTimerStatus(sleepTimeRunning = false)
+
+        LogBuffer.i(LogBuffer.TAG_PLAYBACK, "Sleeping playback for end of episode")
+
+        onSleepEndOfEpisode()
     }
 
-    fun setEndOfChapter() {
+    suspend fun sleepEndOfChapter(onSleepEndOfChapter: suspend () -> Unit) {
+        if (state.isSleepEndOfChapterRunning) {
+            updateSleepTimer {
+                copy(numberOfChaptersLeft = state.numberOfChaptersLeft - 1)
+            }
+            setEndOfChapter()
+        }
+
+        if (state.isSleepEndOfChapterRunning) return
+
+        updateSleepTimerStatus(sleepTimeRunning = false)
+
+        LogBuffer.i(LogBuffer.TAG_PLAYBACK, "Sleeping playback for end of chapters")
+
+        onSleepEndOfChapter()
+    }
+
+    fun cancelTimer() {
+        LogBuffer.i(TAG, "Cleaning automatic sleep timer feature...")
+        updateSleepTimerStatus(sleepTimeRunning = false, sleepAfterChapters = 0, sleepAfterEpisodes = 0)
+        sleepTimerHistory = SleepTimerHistory.None
+    }
+
+    private fun setEndOfEpisodeUuid(uuid: String) {
+        LogBuffer.i(TAG, "Episode $uuid was marked as end of episode")
+        sleepTimerHistory = SleepTimerHistory.AfterEpisode(
+            lastEpisodeUuidAutomaticEnded = uuid,
+            lastFinishedTime = System.currentTimeMillis().milliseconds,
+        )
+    }
+
+    private fun setEndOfChapter() {
         LogBuffer.i(TAG, "End of chapter was reached")
         val time = System.currentTimeMillis().milliseconds
-        lastSleepAfterEndOfChapterTime = time
-        lastTimeSleepTimeHasFinished = time
-        cancelAutomaticSleepAfterTimeRestart()
-        cancelAutomaticSleepOnEpisodeEndRestart()
+        sleepTimerHistory = SleepTimerHistory.AfterChapter(
+            lastSleepAfterEndOfChapterTime = time,
+            lastFinishedTime = time,
+        )
     }
 
-    private fun shouldRestartSleepAfterTime(diffTime: Duration, isSleepTimerRunning: Boolean) = diffTime < MIN_TIME_TO_RESTART_SLEEP_TIMER_IN_MINUTES && lastSleepAfterTime != null && !isSleepTimerRunning
+    private fun shouldRestartSleepAfterTime(diffTime: Duration, isSleepTimerRunning: Boolean): Boolean = diffTime < MIN_TIME_TO_RESTART_SLEEP_TIMER_IN_MINUTES && sleepTimerHistory is SleepTimerHistory.AfterTime && !isSleepTimerRunning
 
     private fun shouldRestartSleepEndOfEpisode(
         diffTime: Duration,
         currentEpisodeUuid: String,
         isSleepEndOfEpisodeRunning: Boolean,
-    ) = diffTime < MIN_TIME_TO_RESTART_SLEEP_TIMER_IN_MINUTES && !lastEpisodeUuidAutomaticEnded.isNullOrEmpty() && currentEpisodeUuid != lastEpisodeUuidAutomaticEnded && !isSleepEndOfEpisodeRunning
+        lastEpisodeUuidAutomaticEnded: String,
+    ) = diffTime < MIN_TIME_TO_RESTART_SLEEP_TIMER_IN_MINUTES && currentEpisodeUuid != lastEpisodeUuidAutomaticEnded && !isSleepEndOfEpisodeRunning
 
-    private fun shouldRestartSleepEndOfChapter(diffTime: Duration, isSleepEndOfChapterRunning: Boolean) = diffTime < MIN_TIME_TO_RESTART_SLEEP_TIMER_IN_MINUTES && !isSleepEndOfChapterRunning && lastSleepAfterEndOfChapterTime != null
+    private fun shouldRestartSleepEndOfChapter(diffTime: Duration, isSleepEndOfChapterRunning: Boolean) = diffTime < MIN_TIME_TO_RESTART_SLEEP_TIMER_IN_MINUTES && !isSleepEndOfChapterRunning
 
-    private fun createAlarm(timeMs: Long): Boolean {
-        val sleepIntent = getSleepIntent()
-        val alarmManager = getAlarmManager()
-        alarmManager.cancel(sleepIntent)
-        return try {
-            LogBuffer.i(TAG, "Starting...")
-            alarmManager.setAndAllowWhileIdle(AlarmManager.RTC_WAKEUP, timeMs, sleepIntent)
-            sleepTimeMs = timeMs
-            lastTimeSleepTimeHasFinished = timeMs.milliseconds
-            true
-        } catch (e: Exception) {
-            LogBuffer.e(LogBuffer.TAG_CRASH, e, "Unable to start sleep timer.")
-            false
-        }
+    private fun updateSleepTimer(update: SleepTimerState.() -> SleepTimerState) {
+        _stateFlow.update { currentState -> currentState.update() }
     }
-
-    fun cancelTimer() {
-        LogBuffer.i(TAG, "Cleaning automatic sleep timer feature...")
-        getAlarmManager().cancel(getSleepIntent())
-        cancelSleepTime()
-        cancelAutomaticSleepAfterTimeRestart()
-        cancelAutomaticSleepOnEpisodeEndRestart()
-        cancelAutomaticSleepOnChapterEndRestart()
-    }
-
-    val isSleepAfterTimerRunning: Boolean
-        get() = System.currentTimeMillis() < (sleepTimeMs ?: -1)
-
-    fun timeLeftInSecs(): Int? {
-        val sleepTimeMs = sleepTimeMs ?: return null
-
-        val timeLeft = sleepTimeMs - System.currentTimeMillis()
-        if (timeLeft < 0) {
-            LogBuffer.i(TAG, "Cancelled because time is up")
-            cancelSleepTime()
-            return null
-        }
-        return (timeLeft / DateUtils.SECOND_IN_MILLIS).toInt()
-    }
-
-    private fun getSleepIntent(): PendingIntent {
-        val intent = Intent(context, SleepTimerReceiver::class.java)
-        return PendingIntent.getBroadcast(context, 234324243, intent, PendingIntent.FLAG_IMMUTABLE)
-    }
-
-    private fun getAlarmManager(): AlarmManager {
-        return context.getSystemService(Context.ALARM_SERVICE) as AlarmManager
-    }
-
-    private fun cancelSleepTime() {
-        sleepTimeMs = null
-    }
-
-    private fun cancelAutomaticSleepAfterTimeRestart() {
-        lastSleepAfterTime = null
-    }
-
-    private fun cancelAutomaticSleepOnEpisodeEndRestart() {
-        lastEpisodeUuidAutomaticEnded = null
-    }
-
-    private fun cancelAutomaticSleepOnChapterEndRestart() {
-        lastSleepAfterEndOfChapterTime = null
-    }
-
-    data class SleepTimerState(
-        val isSleepTimerRunning: Boolean,
-        val isSleepEndOfEpisodeRunning: Boolean,
-        val isSleepEndOfChapterRunning: Boolean,
-        val numberOfEpisodes: Int,
-        val numberOfChapters: Int,
-    )
 }